Add Flask-XML-RPC and Flask-Compress to deb repo
Flask-XML-RPC and Flask-Compress are prerequisites for the Flask backend
for NIPAP. While the Flask backend is still a proof of concept, it is
nearing being good enough to replace the old Twisted backend. Including
these libs as they are not available in usual debian repos and we need
some way of getting them.

Kind of part of #273.